refactor: 누적 맞은 문제수의 변수 타입을 Long에서 Short로 변경

+++ src/main/java/org/choon/careerbee/domain/competition/domain/CompetitionSummary.java
@@ -103,7 +103,7 @@ public class CompetitionSummary extends BaseEntity {
     }
 
     public void updateSummary(TempSummaryInfo summaryInfo) {
-        this.solvedCount = summaryInfo.solvedSum().shortValue();
+        this.solvedCount = summaryInfo.solvedSum();
         this.elapsedTime = summaryInfo.timeSum();
         this.correctRate = summaryInfo.correctRate();
         this.maxContinuousDays = summaryInfo.maxStreak();
