+++ src/main/java/org/choon/careerbee/domain/competition/dto/request/TempSummaryInfo.java
@@ -2,7 +2,7 @@ package org.choon.careerbee.domain.competition.dto.request;
 
 public record TempSummaryInfo(
     Long memberId,
-    Long solvedSum,
+    Short solvedSum,
     Long timeSum,
     Double correctRate,
     Integer maxStreak
